backend/pyproject.toml
[tool.poetry]
name = "scraping-platform"
version = "0.1.0"
description = "Plateforme de scraping intelligent avec LLM"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
motor = "^3.3.1"
httpx = "^0.25.0"
beautifulsoup4 = "^4.12.2"
pandas = "^2.1.2"
xlsxwriter = "^3.1.9"
python-multipart = "^0.0.6"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
aiohttp = "^3.11.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 